chore: version 1.5 SRI calculs et scénarios

+++ src/models/allocation.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, Float
 from src.database import Base
 
 class Allocation(Base):
@@ -9,9 +9,9 @@ class Allocation(Base):
     valo = Column("valo", Integer, nullable=True)
     mouvement = Column("mouvement", Integer, nullable=True)
     sicav = Column("sicav", Integer, nullable=True)
-    perf_sicav_hebdo = Column("perf_sicav_hebdo", Integer, nullable=True)
-    perf_sicav_52 = Column("perf_sicav_52", Integer, nullable=True)
-    volat = Column("volat", Integer, nullable=True)
+    perf_sicav_hebdo = Column("perf_sicav_hebdo", Float, nullable=True)
+    perf_sicav_52 = Column("perf_sicav_52", Float, nullable=True)
+    volat = Column("volat", Float, nullable=True)
     annee = Column("anne", Integer, nullable=True)
     nom = Column("nom", String, nullable=True)
     isin = Column("isin", String, nullable=True)
